# api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Piilot API",
    "version": "1.0.0",
    "description": "API de Piilot. Source de verite du contrat HTTP : le front genere ses types depuis ce document (npm run api:types). Toute route montee doit y figurer, ce que verifie TestSpecOpenAPICouvreToutesLesRoutes."
  },
  "servers": [
    {
      "url": "/",
      "description": "Meme origine que l'application"
    }
  ],
...
  "paths": {
...
    "/api/v1/auth/invitations/{token}/accept": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Accepter une invitation",
        "description": "Cree le compte et ouvre la session (cookies). Bornee par IP.",
        "operationId": "acceptInvitation",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Jeton du lien recu par e-mail."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvitationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Session ouverte",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "404": {
            "description": "INVITATION_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "410": {
            "description": "INVITATION_EXPIRED ou INVITATION_USED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/v1/auth/password/forgot": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Reglage du mot de passe oublie",
        "description": "Dit si l'instance envoie des e-mails.",
        "operationId": "getForgotConfig",
        "responses": {
          "200": {
            "description": "Reglage",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgotConfig"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Mot de passe oublie",
        "description": "Envoie un lien valable une heure si l'adresse est celle d'un compte actif. Repond 202 dans tous les cas : la reponse ne dit pas si l'adresse existe. Bornee par IP et par compte.",
        "operationId": "forgotPassword",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Demande prise en compte"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/v1/auth/password/reset/{token}": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Lien de reinitialisation",
        "description": "Compte vise par le lien. Bornee par IP.",
        "operationId": "getPasswordReset",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Jeton du lien recu par e-mail."
          }
        ],
        "responses": {
          "200": {
            "description": "Compte",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetInfo"
                }
              }
            }
          },
          "410": {
            "description": "RESET_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      },
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Reinitialiser le mot de passe",
        "description": "Change le mot de passe et ferme toutes les sessions du compte. Rend l'adresse, pour pre-remplir la connexion. Bornee par IP.",
        "operationId": "resetPassword",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Jeton du lien recu par e-mail."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mot de passe change",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetInfo"
                }
              }
            }
          },
          "410": {
            "description": "RESET_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "piilot_access",
        "description": "Cookie httpOnly pose par /auth/login. Inaccessible au JavaScript : le navigateur l'envoie seul."
      }
    },
    "responses": {
...
      "ValidationFailed": {
        "description": "Corps invalide. details porte un message par champ fautif.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "RateLimited": {
        "description": "Trop de tentatives. L'en-tete Retry-After donne le delai en secondes.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "Format d'erreur unique de l'API. Le front branche sur code, jamais sur message.",
        "required": [
          "code",
          "message",
          "details"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Code stable, destine au code appelant",
            "examples": [
              "INVALID_CREDENTIALS",
              "SESSION_EXPIRED",
              "FORBIDDEN",
              "VALIDATION_FAILED",
              "RATE_LIMITED",
              "NOT_IMPLEMENTED"
            ]
          },
          "message": {
            "type": "string",
            "description": "Message affichable a l'utilisateur"
          },
          "details": {
            "type": "object",
            "description": "Toujours un objet, jamais null. Porte les erreurs par champ en cas de validation.",
            "additionalProperties": true
          }
        }
      },
      "User": {
        "type": "object",
        "description": "Identite rendue au front. Ne porte ni empreinte de mot de passe ni secret TOTP.",
        "required": [
          "id",
          "email",
          "firstname",
          "lastname",
          "role",
          "avatar_url",
          "permissions",
          "gender",
          "phone",
          "address",
          "postal_code",
          "city",
          "country"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "description": "Code du role. Les roles vivent en base : cette liste est celle des roles systeme, elle peut s'etendre.",
            "examples": [
              "admin",
              "team",
              "client"
            ]
          },
          "avatar_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "gender": {
            "type": "string",
            "description": "Vide quand rien n'est renseigne : un compte n'a pas a se declarer pour exister.",
            "enum": [
              "",
              "male",
              "female",
              "nonbinary"
            ]
          },
          "phone": {
            "type": "string",
            "description": "Numero de telephone. Chaine vide quand il n'y en a pas."
          },
          "address": {
            "type": "string",
            "description": "Voie et complement. Chaine vide quand il n'y en a pas."
          },
          "postal_code": {
            "type": "string",
            "description": "Code postal. Chaine vide quand il n'y en a pas."
          },
          "city": {
            "type": "string",
            "description": "Ville. Chaine vide quand il n'y en a pas."
          },
          "country": {
            "type": "string",
            "description": "Pays. Chaine vide quand il n'y en a pas."
          },
          "permissions": {
            "type": "array",
            "description": "Permissions du role, pour masquer les actions inaccessibles. Confort d'affichage : l'API reste la seule autorite.",
            "items": {
              "type": "string"
            },
            "examples": [
              [
                "projects.read",
                "projects.write"
              ]
            ]
          }
        }
      },
      "SessionResponse": {
        "type": "object",
        "description": "Reponse de login, refresh et me. Les jetons sont dans les cookies, jamais ici.",
        "required": [
          "user"
        ],
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
...
      "AcceptInvitationRequest": {
        "type": "object",
        "description": "Ce que l'invite choisit.",
        "required": [
          "firstname",
          "lastname",
          "password"
        ],
        "properties": {
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "ForgotRequest": {
        "type": "object",
        "description": "Demande de lien de reinitialisation.",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string"
          }
        }
      },
      "ForgotConfig": {
        "type": "object",
        "description": "Reglage de la page mot de passe oublie.",
        "required": [
          "mail_enabled"
        ],
        "properties": {
          "mail_enabled": {
            "type": "boolean"
          }
        }
      },
      "ResetInfo": {
        "type": "object",
        "description": "Compte vise par un lien de reinitialisation.",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string"
          }
        }
      },
      "ResetRequest": {
        "type": "object",
        "description": "Nouveau mot de passe.",
        "required": [
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          }
        }
      }
    }
  }
}